FriscoTap 0.3.0.4
Fixed a weird bug in the RemoteUpdate asynchronous task.  For some reason the FragmentManager kept returning null in onPostExecute when attempting to show the BeersTappedDialog box.  This seemed to pop up out of no where.  The old app uses this method with no issues and has for quite some time.  Moving on for now...